backend/requirements.txt
amqp==5.1.1
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
billiard==4.1.0
celery==5.3.4
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
Django==4.2.5
django-admin-logs==1.0.2
django-cleanup==8.0.0
django-cors-headers==4.2.0
django-debug-toolbar==4.2.0
django-jazzmin==2.6.0
django-modeltranslation==0.18.11
djangorestframework==3.14.0
drf-spectacular==0.26.5
drf-yasg==1.21.7
exceptiongroup==1.1.3
factory-boy==3.3.0
Faker==19.6.2
geographiclib==2.0
geopy==2.4.0
gunicorn==21.2.0
inflection==0.5.1
iniconfig==2.0.0
isort==5.12.0
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
kombu==5.3.2
packaging==23.2
Pillow==10.0.1
pluggy==1.3.0
prompt-toolkit==3.0.39
psycopg2-binary==2.9.8
pytest==7.4.2
pytest-django==4.5.2
python-dateutil==2.8.2
python-decouple==3.8
pytz==2023.3.post1
PyYAML==6.0.1
redis==4.6.0
referencing==0.30.2
rpds-py==0.10.3
ruff==0.1.2
six==1.16.0
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.8.0
tzdata==2023.3
uritemplate==4.1.1
vine==5.0.0
wcwidth==0.2.9
